refactor(characters): extract advancement into Characters.Advancement
Part of #131.

characters.ex sits at the code-health size threshold (~1000 lines),
and the next change — moving progression choice resolution out of the
CLI server — would push it well past. Award application is the start
of a cohesive "advancement" area (awards now, progression choice
resolution next), so it moves to a dedicated Characters.Advancement
module before that lands.

Characters.apply_award/4 remains the public entry point, delegating to
the new module, so callers and the documented API are unchanged. The
implementation is moved verbatim; Advancement reaches back into
Characters only through its public functions (xp_to_next_level,
compute_pending_choice_slots).
